[types] move all content types to it's own module
