+++ polarity/types/content.py
@@ -4,7 +4,9 @@ from time import sleep
 from typing import List
 
 from polarity.types.base import MediaType, MetaMediaType
+from polarity.types.people import Person
 from polarity.types.stream import Stream
+from polarity.utils import normalize_number
 
 
 @dataclass
@@ -12,12 +14,12 @@ class Content(MediaType, metaclass=MetaMediaType):
     title: str
     id: str
     synopsis: str = ""
-    # TODO: better default for date
-    date: datetime = field(default=datetime.now())
-    people: list = field(default_factory=list)
+    date: datetime = field(default=datetime.fromisoformat("1970-01-01"))
+    people: list[Person] = field(default_factory=list)
     genres: List[str] = field(default_factory=list)
-    images: list = field(default_factory=list)
+    images: list[dict] = field(default_factory=list)
     streams: list = field(default_factory=list)
+    # if this value is other than None, don't download the item
     skip_download = None
     output: str = field(init=False, default="")
     _parent: object = field(init=False)
@@ -66,7 +68,7 @@ class ContentContainer(MediaType, metaclass=MetaMediaType):
             self.actors.append(person)
 
     def link_content(self, content: Content) -> None:
-        if self.id == "initial":
+        if self.id == "__polarity_initial":
             # ContentContainer is the initial created by BaseExtractor (self.info),
             # link to content as _parent
             content._parent = self
@@ -113,3 +115,60 @@ class ContentContainer(MediaType, metaclass=MetaMediaType):
         """Sleep until extraction has finished, useful for scripting"""
         while not self._extracted:
             sleep(0.1)
+
+
+@dataclass
+class Series(ContentContainer):
+    synopsis: str = ""
+    year: int = 1970
+    season_count: int = 0
+    episode_count: int = 0
+
+    def __repr__(self) -> str:
+        return f"Series({self.title}, {self.id})"
+
+    def link_content(self, content: Content) -> None:
+        content._series = self
+        return super().link_content(content)
+
+
+@dataclass
+class Season(ContentContainer):
+    number: int = None
+    year: int = 1970
+    images: List[str] = field(default_factory=list)
+    episode_count: int = 0
+    finished: bool = True
+    synopsis: str = ""
+    _series = None
+    _partial = True
+
+    def link_content(self, content: Content):
+        if content not in self.content:
+            content._season = self
+            content._series = self._series
+            self.content.append(content)
+
+
+@dataclass
+class Episode(Content):
+    number: int = 0
+    _series = None
+    _season = None
+
+    @property
+    def short_name(self) -> str:
+        return "%s S%sE%s" % (
+            self._series.title,
+            normalize_number(self._season.number),
+            normalize_number(self.number),
+        )
+
+    @property
+    def content_id(self) -> str:
+        return f"{self._series._extractor.lower()}/episode-{self.id}"
+
+
+@dataclass
+class Movie(Content):
+    pass
