--- polarity/types/movie.py
@@ -1,7 +1,0 @@
-from dataclasses import dataclass
-from polarity.types.generic import Content
-
-
-@dataclass
-class Movie(Content):
-    pass
